requirements.txt
streamlit
pillow
torch
torchvision
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
